tee_inference_worker/requirements.txt
# Core inference server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ML framework (example - adjust based on your model)
numpy==1.26.2
onnxruntime==1.16.3

# Cryptography
cryptography==41.0.7
PyJWT==2.8.0

# HTTP client for GCP metadata
httpx==0.25.2

# System
python-dotenv==1.0.0

